refactor: PomodoroService layer, combined James and Adrians serviec layers
